requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
python-dotenv==1.0.1
cryptography==44.0.0
requests==2.32.3
